+++ pkg/models/deleted-spot-info-model.go
@@ -96,7 +96,7 @@ func (cfg *DeletedSpotInfoConfiguration) InitJob() func() {
 		var budgets []Budget
 		months := map[int64][]time.Time{}
 		badgerBudgets := storage.Open(DbBudgets)
-		err = badgerBudgets.Find(&budgets, badgerhold.Where("Month").Ge(-1))
+		err = badgerBudgets.Find(&budgets, badgerhold.Where("Month").Ge(int64(-1)))
 		if err != nil {
 			fmt.Printf("Q:%s - err:%s", qName, err.Error())
 			return
