+++ pkg/models/program-breaks-model.go
@@ -106,7 +106,7 @@ func (cfg *ProgramBreaksConfiguration) InitJob() func() {
 		var cnl []Channels
 		months := map[int64][]time.Time{}
 		badgerBudgets := storage.Open(DbBudgets)
-		err = badgerBudgets.Find(&budgets, badgerhold.Where("Month").Ge(-1))
+		err = badgerBudgets.Find(&budgets, badgerhold.Where("Month").Ge(int64(-1)))
 		if err != nil {
 			fmt.Printf("Q:%s - err:%s", qName, err.Error())
 			return
